backend/requirements.txt
fastapi>=0.135
uvicorn[standard]>=0.29
sqlalchemy>=2.0
alembic>=1.13
pymysql>=1.1
cryptography>=42
python-jose[cryptography]>=3.3
python-dotenv>=1.0
python-multipart>=0.0.9
pydantic>=2.9
